update per latest internal feedback
On both ads, the end cards look much better, but the type still looks compressed for the “Adobe Creative Cloud”, headline, and CTA text. Anything we can do to make that as sharp as the logos?
The fish lady one, I think the timing needs to be smoothed out still. First frame can hold longer. Cross dissolve can be a little longer smoother, and the text typing on can also be a little slower smoother. Just needs some finesse to feel like the motion of their other stuff.